requirements.txt
hydra-core==1.3.2
omegaconf==2.3.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
numpy<2.0.0
matplotlib==3.10.0
scikit_learn==1.6.1
